Add Update & Delete Mutations to Provider
